fix(ci): Container registry authentication support
Provisioning via volumes does not work as advertised, at least not when
those volumes themselves are provisioned via Docker Compose volumes.

This moves the .docker/config.json content to an environment variable in
the runner's environment that itself is forwarded to the job environment
which in turn employs a pre_build_script to put the content in place.

Passing environment variables the via register subcommand's --env option
would pre-empt the environment list in config.template.toml, so that has
been moved to the register service's command line invocation.

The DOCKER_CONFIG_JSON variable is meant to be set via a .env file as it
contains registry authentication credentials that should be kept private
and out of the repository.

Re #5