fix: Use migrated images from Docker Hub for build jobs
The build results are pushed to a "temporary" registry where the test
job can pick them up.  The deploy job will also look there for images
that can be deployed but still deploys to the same location as before
(which just happens to be the same as the "temporary" registry).

The migrate job still pushes to the "temporary" registry as well.

Re #22