refactor: Move tagging/pushing logic into a script
The script handles tagging and pushing of both the devuan and migrated
images to their target container registry.

To enable its use, the "testable" images have been renamed.  Because a
unique build identifier is used that does not conflict with the set of
tags to be pushed to the target container registry this is okay.  Even
if source and target repositories are the same, tags won't conflict.

The script to publish images only considers a BUILD_VARIANT.  This is
independent of the Debian image's IMAGE_VARIANT used to migrate.